requirements.txt
black==22.10.0
Django==5.0.7
djangorestframework==3.15.2
django-timestampedmodel==0.1.0
pre-commit==3.8.0
requests==2.32.3
